GamesController endpoints return ActionResult

+++ LoupGarou/Controllers/GamesController.cs
@@ -18,33 +18,41 @@ namespace LoupGarou.Controllers
 
     // GET: api/<GamesController>
     [HttpGet]
-    public async Task<IEnumerable<Game>> Get()
+    public async Task<ActionResult<IEnumerable<Game>>> Get()
     {
-      var allGame = await gameService.GetAllGames();
-      return allGame;
+      var allGames = await gameService.GetAllGames();
+      if(allGames == null) return NoContent();
+      return Ok(allGames);
     }
 
     // GET api/<GamesController>/5
     [HttpGet("{id}")]
-    public async Task<Game> Get(string id)
+    public async Task<ActionResult<Game>> Get(string id)
     {
       var game = await gameService.GetGame(id);
-      return game;
+      if(game == null) return NotFound();
+      return Ok(game);
     }
 
     // POST api/<GamesController>
     [HttpPost]
-    public async Task<string> Post([FromBody] int numberOfPlayers)
+    public async Task<ActionResult<string>> Post([FromBody] int numberOfPlayers)
     {
-      string id = await gameService.CreateGame(numberOfPlayers);
-      return id;
+      Game game = await gameService.CreateGame(numberOfPlayers);
+      var baseUrl = $"{HttpContext.Request.Scheme}://{HttpContext.Request.Host.ToUriComponent()}";
+      var getUrl = baseUrl + "/api/Games/" +  game.GameId;
+      return Created( getUrl, game);
     }
 
     // DELETE api/<GamesController>/5
     [HttpDelete("{id}")]
-    public async Task Delete(string id)
+    public async Task<ActionResult> Delete(string id)
     {
-       await gameService.DeleteGame(id);
+      var game = await gameService.GetGame(id);
+      if(game == null) return NotFound();
+
+      await gameService.DeleteGame(id);
+      return NoContent();
     }
   }
 }
